+++ internal-api/src/main/java/datadog/trace/profiling/Profiler.java
@@ -1,4 +1,4 @@
-package datadog.trace.api.profiling;
+package datadog.trace.profiling;
 
 import lombok.extern.slf4j.Slf4j;
