+++ internal-api/src/main/java/datadog/trace/profiling/Session.java
@@ -1,4 +1,4 @@
-package datadog.trace.api.profiling;
+package datadog.trace.profiling;
 
 import java.io.Closeable;
